Send notifs to user when actu or events

+++ docker-rails.sh
@@ -1,4 +1,11 @@
-#!/bin/sh
+#!/bin/bash
+set -e
 
-rm -f tmp/pids/server.pid
-bin/rails server -b 0.0.0.0 -p 3000
+if [ "$RAILS_ENV" = "production" ]; then
+  echo "===> Starting Unicorn in production"
+  bundle exec unicorn -c /app/config/unicorn.rb -E production
+else
+  echo "===> Starting Rails server in $RAILS_ENV"
+  rm -f tmp/pids/server.pid
+  bin/rails server -b 0.0.0.0 -p 3000
+fi
